fix: prevent bootstrap XP threshold overflow (#194)

+++ src/Elyndor.Contracts/World/WorldContracts.cs
@@ -69,7 +69,7 @@ public sealed record BootstrapCharacterResponse(
     string ClassId,
     int Level,
     long Experience,
-    int XpToNextLevel,
+    long XpToNextLevel,
     long Gold,
     string PrimaryAttribute,
     string ClassProfileVersion,
